KAFKA-14683: Migrate WorkerSinkTaskTest to Mockito (2/3) (#15313)
Reviewers: Greg Harris <greg.harris@aiven.io>
